src/githubai/requirements.txt
requests
openai
pyyaml
pygithub